+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- KeyCloak method to retrieve a access_token to be used as Bearer authentication
 
 ## [1.6.9](https://github.com/kb-dk/kb-util/tree/kb-util-1.6.9) - 2025-03-03
 ### Fixed